Store HomeKit values, and scale when setting LED
This prevents rounding errors when retrieving previously set values.

+++ src/lightstripmodel.h
@@ -11,7 +11,7 @@
 class LightStrip {
     public:
         bool isOn;
-        byte hue;
+        int hue;
         byte saturation;
         byte brightness;
         String name;
